[main] configure transactional

+++ src/main/java/com/gate/demo/service/impl/UserServiceImpl.java
@@ -8,7 +8,10 @@ import com.gate.demo.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -25,6 +28,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserResponseDto createUser(UserRequestDto request) {
+
+        if (userRepository.existsByEmail(request.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Email already exists");
+        }
 
         User user = new User();
         user.setId(UUID.randomUUID());
@@ -58,7 +65,7 @@ public class UserServiceImpl implements UserService {
 
     public UserResponseDto getUser(UUID id) {
         User usr = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
 
         UserResponseDto response = new UserResponseDto();
         response.setId(usr.getId());
@@ -69,24 +76,22 @@ public class UserServiceImpl implements UserService {
         return response;
     }
 
-    public UserResponseDto updateUser(UUID id, UserRequestDto requestDto) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+    public ResponseEntity<String> updateUser(UserRequestDto requestDto) {
+        User user = userRepository.findById(requestDto.getId())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
         user.setName(requestDto.getName());
         user.setEmail(requestDto.getEmail());
         userRepository.save(user);
 
-        UserResponseDto response = new UserResponseDto();
-        response.setId(user.getId());
-        response.setName(user.getName());
-        response.setEmail(user.getEmail());
-        response.setCreated_at(user.getCreatedAt());
-
-        return response;
+        return new ResponseEntity<>("User updated successfully", HttpStatus.OK);
     }
 
-    public void deleteUser(UUID id) {
+    public ResponseEntity<String> deleteUser(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
+        }
         userRepository.deleteById(id);
+        return new ResponseEntity<>("User deleted successfully", HttpStatus.OK);
     }
 }
